+++ pkg/audit/verifier_test.go
@@ -5,7 +5,6 @@ package audit_test
 
 import (
 	"context"
-	"fmt"
 	"math/rand"
 	"testing"
 	"time"
@@ -24,7 +23,6 @@ import (
 	"storj.io/storj/pkg/peertls/tlsopts"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/pkg/transport"
-	"storj.io/storj/storagenode"
 	"storj.io/storj/uplink"
 )
 
@@ -633,29 +631,4 @@ func TestVerifierModifiedSegment(t *testing.T) {
 		require.True(t, audit.ErrSegmentDeleted.Has(err))
 		assert.Empty(t, report)
 	})
-}
-
-func getStorageNode(planet *testplanet.Planet, nodeID storj.NodeID) *storagenode.Peer {
-	for _, node := range planet.StorageNodes {
-		if node.ID() == nodeID {
-			return node
-		}
-	}
-	return nil
-}
-
-func stopStorageNode(ctx context.Context, planet *testplanet.Planet, nodeID storj.NodeID) error {
-	node := getStorageNode(planet, nodeID)
-	if node == nil {
-		return fmt.Errorf("no such node: %s", nodeID.String())
-	}
-
-	err := planet.StopPeer(node)
-	if err != nil {
-		return err
-	}
-
-	// mark stopped node as offline in overlay cache
-	_, err = planet.Satellites[0].Overlay.Service.UpdateUptime(ctx, nodeID, false)
-	return err
 }
